services/agents/pyproject.toml
[project]
name = "agents-service"
version = "0.1.0"
description = "Python agent service for mee6"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "anthropic",
    "pydantic",
    "pydantic-settings",
    "playwright",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
